refactor: extract createReviewDatetimeCycle logic
